refactor(tools): centralize sandbox filesystem access
Entire-Checkpoint: 291f022d0952
