+++ crates/tools/src/sandbox/file_system.rs
@@ -1,0 +1,615 @@
+//! Session-scoped filesystem access inside sandboxes.
+//!
+//! This provides a first-class file service for sandboxed tools so callers do
+//! not need to assemble shell snippets or juggle `(backend, sandbox_id)` pairs
+//! directly. The current implementation still uses `Sandbox::exec` under the
+//! hood, but the shell transport is now hidden behind a stable Rust interface.
+
+use {
+    async_trait::async_trait,
+    base64::{Engine as _, engine::general_purpose::STANDARD as BASE64},
+    serde_json::{Value, json},
+    std::{path::PathBuf, sync::Arc, time::Duration},
+};
+
+use crate::{
+    Result,
+    error::Error,
+    exec::ExecOpts,
+    sandbox::{Sandbox, SandboxId, SandboxRouter},
+};
+
+/// Maximum file size Write/Edit/MultiEdit can send into a sandbox in a
+/// single call. Base64 expands by ~33%, so 512 KB raw becomes ~683 KB
+/// of shell arg, comfortably under typical `ARG_MAX` limits.
+pub const MAX_SANDBOX_WRITE_BYTES: usize = 512 * 1024;
+
+const DEFAULT_SANDBOX_TIMEOUT: Duration = Duration::from_secs(30);
+const DEFAULT_SANDBOX_OUTPUT_BYTES: usize = 32 * 1024 * 1024;
+
+// Exit codes used by the bridge scripts to encode typed errors.
+const EXIT_NOT_FOUND: i32 = 10;
+const EXIT_PERMISSION_DENIED: i32 = 11;
+const EXIT_NOT_REGULAR_FILE: i32 = 12;
+const EXIT_TOO_LARGE: i32 = 13;
+const EXIT_SYMLINK: i32 = 14;
+const EXIT_PARENT_MISSING: i32 = 20;
+
+fn default_opts() -> ExecOpts {
+    ExecOpts {
+        timeout: DEFAULT_SANDBOX_TIMEOUT,
+        max_output_bytes: DEFAULT_SANDBOX_OUTPUT_BYTES,
+        working_dir: Some(PathBuf::from("/home/sandbox")),
+        env: Vec::new(),
+    }
+}
+
+/// Escape a string for safe use inside single quotes in a POSIX shell.
+#[must_use]
+pub fn shell_single_quote(input: &str) -> String {
+    format!("'{}'", input.replace('\'', "'\\''"))
+}
+
+/// Outcome of a sandbox read call.
+#[derive(Debug)]
+pub enum SandboxReadResult {
+    Ok(Vec<u8>),
+    NotFound,
+    PermissionDenied,
+    NotRegularFile,
+    TooLarge(u64),
+}
+
+impl SandboxReadResult {
+    /// Convert a non-`Ok` variant into the typed JSON payload the fs
+    /// tools return. `None` for `Ok` (caller handles success).
+    #[must_use]
+    pub fn into_typed_payload(self, file_path: &str, max_bytes: u64) -> Option<Value> {
+        match self {
+            Self::Ok(_) => None,
+            Self::NotFound => Some(json!({
+                "kind": "not_found",
+                "file_path": file_path,
+                "error": "file does not exist",
+                "detail": "",
+            })),
+            Self::PermissionDenied => Some(json!({
+                "kind": "permission_denied",
+                "file_path": file_path,
+                "error": "insufficient permissions to access file",
+                "detail": "",
+            })),
+            Self::NotRegularFile => Some(json!({
+                "kind": "not_regular_file",
+                "file_path": file_path,
+                "error": "path is not a regular file",
+                "detail": "",
+            })),
+            Self::TooLarge(size) => Some(json!({
+                "kind": "too_large",
+                "file_path": file_path,
+                "error": format!(
+                    "file is too large ({:.1} MB) — maximum is {:.0} MB",
+                    size as f64 / (1024.0 * 1024.0),
+                    max_bytes as f64 / (1024.0 * 1024.0),
+                ),
+                "bytes": size,
+                "max_bytes": max_bytes,
+            })),
+        }
+    }
+}
+
+/// Output-mode discriminator for sandbox grep.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SandboxGrepMode {
+    Content,
+    FilesWithMatches,
+    Count,
+}
+
+/// Options for sandbox grep.
+#[derive(Debug, Clone)]
+pub struct SandboxGrepOptions {
+    pub pattern: String,
+    pub path: String,
+    pub mode: SandboxGrepMode,
+    pub case_insensitive: bool,
+    pub include_globs: Vec<String>,
+    pub offset: usize,
+    pub head_limit: Option<usize>,
+    pub match_cap: Option<usize>,
+}
+
+/// Abstract filesystem access for a prepared sandbox session.
+#[async_trait]
+pub trait SandboxFileSystem: Send + Sync {
+    async fn read_file(&self, file_path: &str, max_bytes: u64) -> Result<SandboxReadResult>;
+
+    async fn write_file(&self, file_path: &str, content: &[u8]) -> Result<Option<Value>>;
+
+    async fn list_files(&self, root: &str) -> Result<Vec<String>>;
+
+    async fn grep(&self, opts: SandboxGrepOptions) -> Result<Value>;
+}
+
+/// Command-based [`SandboxFileSystem`] implementation backed by `Sandbox::exec`.
+pub struct CommandSandboxFileSystem {
+    backend: Arc<dyn Sandbox>,
+    id: SandboxId,
+}
+
+impl CommandSandboxFileSystem {
+    #[must_use]
+    pub fn new(backend: Arc<dyn Sandbox>, id: SandboxId) -> Self {
+        Self { backend, id }
+    }
+}
+
+/// Prepare a session sandbox and return a file service for it.
+pub async fn sandbox_file_system_for_session(
+    router: &SandboxRouter,
+    session_key: &str,
+) -> Result<Arc<dyn SandboxFileSystem>> {
+    let id = router.sandbox_id_for(session_key);
+    let image = router.resolve_image(session_key, None).await;
+    let backend = Arc::clone(router.backend());
+    backend.ensure_ready(&id, Some(&image)).await?;
+    Ok(Arc::new(CommandSandboxFileSystem::new(backend, id)))
+}
+
+#[async_trait]
+impl SandboxFileSystem for CommandSandboxFileSystem {
+    async fn read_file(&self, file_path: &str, max_bytes: u64) -> Result<SandboxReadResult> {
+        let quoted = shell_single_quote(file_path);
+        let script = format!(
+            "path={quoted}; max={max_bytes}; \
+             if [ ! -e \"$path\" ]; then exit {EXIT_NOT_FOUND}; fi; \
+             if [ ! -r \"$path\" ]; then exit {EXIT_PERMISSION_DENIED}; fi; \
+             if [ ! -f \"$path\" ]; then exit {EXIT_NOT_REGULAR_FILE}; fi; \
+             size=$(wc -c < \"$path\"); \
+             if [ \"$size\" -gt \"$max\" ]; then echo \"$size\" >&2; exit {EXIT_TOO_LARGE}; fi; \
+             base64 < \"$path\" | tr -d '\\n'"
+        );
+
+        let result = self
+            .backend
+            .exec(&self.id, &script, &default_opts())
+            .await?;
+        match result.exit_code {
+            0 => {
+                let bytes = BASE64.decode(result.stdout.trim()).map_err(|e| {
+                    Error::message(format!(
+                        "failed to decode sandbox read for '{file_path}': {e}"
+                    ))
+                })?;
+                Ok(SandboxReadResult::Ok(bytes))
+            },
+            EXIT_NOT_FOUND => Ok(SandboxReadResult::NotFound),
+            EXIT_PERMISSION_DENIED => Ok(SandboxReadResult::PermissionDenied),
+            EXIT_NOT_REGULAR_FILE => Ok(SandboxReadResult::NotRegularFile),
+            EXIT_TOO_LARGE => {
+                let size = result.stderr.trim().parse::<u64>().unwrap_or(0);
+                Ok(SandboxReadResult::TooLarge(size))
+            },
+            other => {
+                let detail = if result.stderr.trim().is_empty() {
+                    format!("sandbox read exited with code {other}")
+                } else {
+                    result.stderr.trim().to_string()
+                };
+                Err(Error::message(format!(
+                    "sandbox read of '{file_path}' failed: {detail}"
+                )))
+            },
+        }
+    }
+
+    async fn write_file(&self, file_path: &str, content: &[u8]) -> Result<Option<Value>> {
+        if content.len() > MAX_SANDBOX_WRITE_BYTES {
+            return Err(Error::message(format!(
+                "sandbox Write is limited to {} KB per call (got {:.1} KB); \
+                 larger writes will ship in a follow-up that chunks content",
+                MAX_SANDBOX_WRITE_BYTES / 1024,
+                content.len() as f64 / 1024.0,
+            )));
+        }
+
+        let encoded = BASE64.encode(content);
+        let quoted_path = shell_single_quote(file_path);
+        let script = format!(
+            "path={quoted_path}; \
+             parent=$(dirname \"$path\"); \
+             if [ ! -d \"$parent\" ]; then exit {EXIT_PARENT_MISSING}; fi; \
+             if [ -L \"$path\" ]; then exit {EXIT_SYMLINK}; fi; \
+             tmp=\"$path.moltis.$$\"; \
+             if ! printf '%s' '{encoded}' | base64 -d > \"$tmp\"; then rm -f \"$tmp\"; exit 1; fi; \
+             sync \"$tmp\" 2>/dev/null || sync; \
+             if [ -L \"$path\" ]; then rm -f \"$tmp\"; exit {EXIT_SYMLINK}; fi; \
+             mv \"$tmp\" \"$path\""
+        );
+
+        let result = self
+            .backend
+            .exec(&self.id, &script, &default_opts())
+            .await?;
+        match result.exit_code {
+            0 => Ok(None),
+            EXIT_PARENT_MISSING => Err(Error::message(format!(
+                "cannot resolve parent of '{file_path}': directory does not exist in sandbox"
+            ))),
+            EXIT_SYMLINK => Ok(Some(json!({
+                "kind": "path_denied",
+                "file_path": file_path,
+                "error": "target is a symbolic link; refusing to follow",
+                "detail": "sandbox Write rejects symlinks",
+            }))),
+            other => {
+                let detail = if result.stderr.trim().is_empty() {
+                    format!("sandbox write exited with code {other}")
+                } else {
+                    result.stderr.trim().to_string()
+                };
+                Err(Error::message(format!(
+                    "sandbox write of '{file_path}' failed: {detail}"
+                )))
+            },
+        }
+    }
+
+    async fn list_files(&self, root: &str) -> Result<Vec<String>> {
+        let quoted = shell_single_quote(root);
+        let script = format!("find {quoted} -type f 2>/dev/null");
+        let result = self
+            .backend
+            .exec(&self.id, &script, &default_opts())
+            .await?;
+        if result.exit_code != 0 && result.stdout.trim().is_empty() {
+            let detail = if result.stderr.trim().is_empty() {
+                format!("find exited with code {}", result.exit_code)
+            } else {
+                result.stderr.trim().to_string()
+            };
+            return Err(Error::message(format!(
+                "sandbox list_files '{root}' failed: {detail}"
+            )));
+        }
+        Ok(result
+            .stdout
+            .lines()
+            .map(str::trim)
+            .filter(|line| !line.is_empty())
+            .map(ToOwned::to_owned)
+            .collect())
+    }
+
+    async fn grep(&self, opts: SandboxGrepOptions) -> Result<Value> {
+        let pattern_q = shell_single_quote(&opts.pattern);
+        let path_q = shell_single_quote(&opts.path);
+        let mut flags: Vec<&str> = vec!["-r", "-P"];
+        if opts.case_insensitive {
+            flags.push("-i");
+        }
+        match opts.mode {
+            SandboxGrepMode::Content => {
+                flags.push("-n");
+                flags.push("-H");
+            },
+            SandboxGrepMode::FilesWithMatches => {
+                flags.push("-l");
+            },
+            SandboxGrepMode::Count => {
+                flags.push("-c");
+                flags.push("-H");
+            },
+        }
+        let include_args = if opts.include_globs.is_empty() {
+            String::new()
+        } else {
+            opts.include_globs
+                .iter()
+                .map(|glob| format!("--include={}", shell_single_quote(glob)))
+                .collect::<Vec<_>>()
+                .join(" ")
+        };
+        let flags_str = flags.join(" ");
+        let flags_str_ere = flags_str.replace("-P", "-E");
+        let script = format!(
+            "grep {flags_str} {include_args} -- {pattern_q} {path_q} 2>/dev/null; \
+             rc=$?; \
+             if [ $rc -eq 2 ]; then \
+               grep {flags_str_ere} {include_args} -- {pattern_q} {path_q} 2>/dev/null; \
+               rc=$?; \
+             fi; \
+             if [ $rc -eq 1 ]; then exit 0; else exit $rc; fi"
+        );
+        let result = self
+            .backend
+            .exec(&self.id, &script, &default_opts())
+            .await?;
+        if result.exit_code != 0 {
+            let detail = if result.stderr.trim().is_empty() {
+                format!("grep exited with code {}", result.exit_code)
+            } else {
+                result.stderr.trim().to_string()
+            };
+            return Err(Error::message(format!("sandbox grep failed: {detail}")));
+        }
+
+        let lines: Vec<&str> = result
+            .stdout
+            .lines()
+            .map(str::trim_end)
+            .filter(|line| !line.is_empty())
+            .collect();
+
+        match opts.mode {
+            SandboxGrepMode::FilesWithMatches => {
+                let files = lines
+                    .iter()
+                    .map(|line| line.to_string())
+                    .collect::<Vec<_>>();
+                let (files, truncated) = apply_head_offset(files, opts.offset, opts.head_limit);
+                Ok(json!({
+                    "mode": "files_with_matches",
+                    "files": files,
+                    "truncated": truncated,
+                }))
+            },
+            SandboxGrepMode::Count => {
+                let mut counts = Vec::new();
+                for line in &lines {
+                    if let Some((path, count_str)) = line.rsplit_once(':')
+                        && let Ok(count) = count_str.parse::<usize>()
+                        && count > 0
+                    {
+                        counts.push(json!({
+                            "path": path,
+                            "count": count,
+                        }));
+                    }
+                }
+                let (counts, truncated) = apply_head_offset(counts, opts.offset, opts.head_limit);
+                Ok(json!({
+                    "mode": "count",
+                    "counts": counts,
+                    "truncated": truncated,
+                }))
+            },
+            SandboxGrepMode::Content => {
+                let mut matches = Vec::new();
+                for line in &lines {
+                    let mut parts = line.splitn(3, ':');
+                    let (Some(path), Some(lineno_str), Some(text)) =
+                        (parts.next(), parts.next(), parts.next())
+                    else {
+                        continue;
+                    };
+                    let Ok(lineno) = lineno_str.parse::<usize>() else {
+                        continue;
+                    };
+                    matches.push(json!({
+                        "path": path,
+                        "line": lineno,
+                        "match": text,
+                        "block": vec![format!("{lineno}:{text}")],
+                    }));
+                }
+                let (matches, cap_truncated) = apply_match_cap(matches, opts.match_cap);
+                let (matches, page_truncated) =
+                    apply_head_offset(matches, opts.offset, opts.head_limit);
+                Ok(json!({
+                    "mode": "content",
+                    "matches": matches,
+                    "truncated": cap_truncated || page_truncated,
+                }))
+            },
+        }
+    }
+}
+
+fn apply_match_cap<T>(mut rows: Vec<T>, match_cap: Option<usize>) -> (Vec<T>, bool) {
+    match match_cap {
+        Some(limit) if rows.len() > limit => {
+            rows.truncate(limit);
+            (rows, true)
+        },
+        _ => (rows, false),
+    }
+}
+
+fn apply_head_offset<T: Clone>(
+    rows: Vec<T>,
+    offset: usize,
+    head_limit: Option<usize>,
+) -> (Vec<T>, bool) {
+    let total = rows.len();
+    let start = offset.min(total);
+    let slice = &rows[start..];
+    let (capped, truncated) = match head_limit {
+        Some(limit) if slice.len() > limit => (&slice[..limit], true),
+        _ => (slice, false),
+    };
+    (capped.to_vec(), truncated)
+}
+
+#[allow(clippy::unwrap_used, clippy::expect_used)]
+#[cfg(test)]
+pub(crate) mod test_helpers {
+    use {
+        super::*,
+        crate::{exec::ExecResult, sandbox::types::BuildImageResult},
+        std::sync::Mutex,
+    };
+
+    pub struct MockSandbox {
+        pub calls: Mutex<Vec<String>>,
+        pub responses: Mutex<Vec<ExecResult>>,
+    }
+
+    impl MockSandbox {
+        pub fn new(responses: Vec<ExecResult>) -> Arc<Self> {
+            Arc::new(Self {
+                calls: Mutex::new(Vec::new()),
+                responses: Mutex::new(responses),
+            })
+        }
+
+        pub fn last_command(&self) -> Option<String> {
+            self.calls.lock().unwrap().last().cloned()
+        }
+    }
+
+    #[async_trait]
+    impl Sandbox for MockSandbox {
+        fn backend_name(&self) -> &'static str {
+            "mock"
+        }
+
+        fn is_real(&self) -> bool {
+            true
+        }
+
+        async fn ensure_ready(&self, _id: &SandboxId, _image_override: Option<&str>) -> Result<()> {
+            Ok(())
+        }
+
+        async fn exec(
+            &self,
+            _id: &SandboxId,
+            command: &str,
+            _opts: &ExecOpts,
+        ) -> Result<ExecResult> {
+            self.calls.lock().unwrap().push(command.to_string());
+            let mut responses = self.responses.lock().unwrap();
+            if responses.is_empty() {
+                Ok(ExecResult {
+                    stdout: String::new(),
+                    stderr: String::new(),
+                    exit_code: 0,
+                })
+            } else {
+                Ok(responses.remove(0))
+            }
+        }
+
+        async fn cleanup(&self, _id: &SandboxId) -> Result<()> {
+            Ok(())
+        }
+
+        async fn build_image(
+            &self,
+            _base: &str,
+            _packages: &[String],
+        ) -> Result<Option<BuildImageResult>> {
+            Ok(None)
+        }
+    }
+}
+
+#[allow(clippy::unwrap_used, clippy::expect_used)]
+#[cfg(test)]
+mod tests {
+    use {
+        super::{test_helpers::MockSandbox, *},
+        crate::{exec::ExecResult, sandbox::types::SandboxScope},
+    };
+
+    fn test_id() -> SandboxId {
+        SandboxId {
+            scope: SandboxScope::Session,
+            key: "test".to_string(),
+        }
+    }
+
+    #[tokio::test]
+    async fn read_file_decodes_base64() {
+        let encoded = BASE64.encode(b"hello sandbox");
+        let mock = MockSandbox::new(vec![ExecResult {
+            stdout: encoded,
+            stderr: String::new(),
+            exit_code: 0,
+        }]);
+        let fs = CommandSandboxFileSystem::new(mock.clone(), test_id());
+
+        let result = fs.read_file("/data/x.txt", 1024).await.unwrap();
+        match result {
+            SandboxReadResult::Ok(bytes) => assert_eq!(bytes, b"hello sandbox"),
+            other => panic!("expected Ok, got {other:?}"),
+        }
+        assert!(mock.last_command().unwrap().contains("/data/x.txt"));
+    }
+
+    #[tokio::test]
+    async fn read_file_maps_too_large() {
+        let mock = MockSandbox::new(vec![ExecResult {
+            stdout: String::new(),
+            stderr: "12345\n".to_string(),
+            exit_code: EXIT_TOO_LARGE,
+        }]);
+        let fs = CommandSandboxFileSystem::new(mock, test_id());
+
+        let result = fs.read_file("/big", 100).await.unwrap();
+        assert!(matches!(result, SandboxReadResult::TooLarge(12345)));
+    }
+
+    #[tokio::test]
+    async fn write_file_encodes_content() {
+        let mock = MockSandbox::new(vec![ExecResult {
+            stdout: String::new(),
+            stderr: String::new(),
+            exit_code: 0,
+        }]);
+        let fs = CommandSandboxFileSystem::new(mock.clone(), test_id());
+
+        let result = fs.write_file("/data/out.txt", b"abc").await.unwrap();
+        assert!(result.is_none());
+        let cmd = mock.last_command().unwrap();
+        assert!(cmd.contains("/data/out.txt"));
+        assert!(cmd.contains(&BASE64.encode(b"abc")));
+        assert!(cmd.contains("sync \"$tmp\""));
+    }
+
+    #[tokio::test]
+    async fn list_files_reads_find_output() {
+        let mock = MockSandbox::new(vec![ExecResult {
+            stdout: "/data/a.rs\n/data/b.rs\n".to_string(),
+            stderr: String::new(),
+            exit_code: 0,
+        }]);
+        let fs = CommandSandboxFileSystem::new(mock, test_id());
+
+        let files = fs.list_files("/data").await.unwrap();
+        assert_eq!(files, vec!["/data/a.rs", "/data/b.rs"]);
+    }
+
+    #[tokio::test]
+    async fn grep_content_applies_paging() {
+        let mock = MockSandbox::new(vec![ExecResult {
+            stdout: "/data/lib.rs:3:fn alpha()\n/data/lib.rs:9:fn beta()\n".to_string(),
+            stderr: String::new(),
+            exit_code: 0,
+        }]);
+        let fs = CommandSandboxFileSystem::new(mock, test_id());
+
+        let value = fs
+            .grep(SandboxGrepOptions {
+                pattern: "fn".to_string(),
+                path: "/data".to_string(),
+                mode: SandboxGrepMode::Content,
+                case_insensitive: false,
+                include_globs: Vec::new(),
+                offset: 1,
+                head_limit: Some(1),
+                match_cap: None,
+            })
+            .await
+            .unwrap();
+
+        assert_eq!(value["mode"], "content");
+        assert_eq!(value["truncated"], false);
+        let matches = value["matches"].as_array().unwrap();
+        assert_eq!(matches.len(), 1);
+        assert_eq!(matches[0]["line"], 9);
+    }
+}
