+++ crates/tools/src/sandbox/mod.rs
@@ -6,6 +6,7 @@
 pub(crate) mod apple;
 pub(crate) mod containers;
 pub(crate) mod docker;
+pub(crate) mod file_system;
 pub(crate) mod host;
 pub(crate) mod paths;
 pub(crate) mod platform;
